+++ 2023/SpringBoot/src/main/java/io/github/nibiroo/domain/repository/CustomersRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface CustomersRepository extends JpaRepository<Customer, Integer> {
+public interface CustomersRepository extends JpaRepository<Customer, Long> {
 
     Customer findByName(String name);
 
@@ -17,10 +17,10 @@ public interface CustomersRepository extends JpaRepository<Customer, Integer> {
     Boolean existsByName(String name);
 
     @Query(value = "select c from Customer c where c.id = :idParam")
-    List<Customer> findSomethingIDJPQL(@Param("idParam") Integer idParam);
+    List<Customer> findSomethingIDJPQL(@Param("idParam") Long idParam);
 
     @Query(value = "select * from customer c where c.id = :idParam", nativeQuery = true)
-    List<Customer> findSomethingIDSQL(@Param("idParam") Integer idParam);
+    List<Customer> findSomethingIDSQL(@Param("idParam") Long idParam);
 
     @Query(value = "select * from customer c where c.name like :nameParam%", nativeQuery = true)
     List<Customer> findSomethingStartsNameSQL(@Param("nameParam") String nameParam);
@@ -30,5 +30,5 @@ public interface CustomersRepository extends JpaRepository<Customer, Integer> {
     void deleteByName (String name);
 
     @Query(value = "select c from Customer c left join fetch c.invoices where c.id = :idParam")
-    Customer findCustomerFetchInvoices(@Param("idParam") Integer idParam);
+    Customer findCustomerFetchInvoices(@Param("idParam") Long idParam);
 }
